Extract `printFunctionType`
