+++ src/language-js/print/type-annotation.js
@@ -5,13 +5,16 @@ const {
   builders: { concat, group, join, line, softline, indent, align, ifBreak },
 } = require("../../document");
 const pathNeedsParens = require("../needs-parens");
+const { locStart } = require("../loc");
 const {
   isFlowAnnotationComment,
   isSimpleType,
   isObjectType,
   hasLeadingOwnLineComment,
+  isObjectTypePropertyAFunction,
 } = require("../utils");
 const { printAssignmentRight } = require("./assignment");
+const { printFunctionParameters } = require("./function-parameters");
 
 function printTypeAnnotation(path, options, print) {
   const node = path.getValue();
@@ -228,11 +231,82 @@ function printUnionType(path, options, print) {
   return group(shouldIndent ? indent(code) : code);
 }
 
+function printFunctionType(path, options, print) {
+  const n = path.getValue();
+  const parts = [];
+  // FunctionTypeAnnotation is ambiguous:
+  // declare function foo(a: B): void; OR
+  // var A: (a: B) => void;
+  const parent = path.getParentNode(0);
+  const parentParent = path.getParentNode(1);
+  const parentParentParent = path.getParentNode(2);
+  let isArrowFunctionTypeAnnotation =
+    n.type === "TSFunctionType" ||
+    !(
+      ((parent.type === "ObjectTypeProperty" ||
+        parent.type === "ObjectTypeInternalSlot") &&
+        !parent.variance &&
+        !parent.optional &&
+        locStart(parent) === locStart(n)) ||
+      parent.type === "ObjectTypeCallProperty" ||
+      (parentParentParent && parentParentParent.type === "DeclareFunction")
+    );
+
+  let needsColon =
+    isArrowFunctionTypeAnnotation &&
+    (parent.type === "TypeAnnotation" || parent.type === "TSTypeAnnotation");
+
+  // Sadly we can't put it inside of FastPath::needsColon because we are
+  // printing ":" as part of the expression and it would put parenthesis
+  // around :(
+  const needsParens =
+    needsColon &&
+    isArrowFunctionTypeAnnotation &&
+    (parent.type === "TypeAnnotation" || parent.type === "TSTypeAnnotation") &&
+    parentParent.type === "ArrowFunctionExpression";
+
+  if (isObjectTypePropertyAFunction(parent)) {
+    isArrowFunctionTypeAnnotation = true;
+    needsColon = true;
+  }
+
+  if (needsParens) {
+    parts.push("(");
+  }
+
+  parts.push(
+    printFunctionParameters(
+      path,
+      print,
+      options,
+      /* expandArg */ false,
+      /* printTypeParams */ true
+    )
+  );
+
+  // The returnType is not wrapped in a TypeAnnotation, so the colon
+  // needs to be added separately.
+  if (n.returnType || n.predicate || n.typeAnnotation) {
+    parts.push(
+      isArrowFunctionTypeAnnotation ? " => " : ": ",
+      path.call(print, "returnType"),
+      path.call(print, "predicate"),
+      path.call(print, "typeAnnotation")
+    );
+  }
+  if (needsParens) {
+    parts.push(")");
+  }
+
+  return group(concat(parts));
+}
+
 module.exports = {
   printOpaqueType,
   printTypeAlias,
   printTypeAnnotation,
   printIntersectionType,
   printUnionType,
+  printFunctionType,
   shouldHugType,
 };
